requirements-selin.txt
about-time==4.2.1
absl-py==2.3.1
accelerate==1.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
albucore==0.0.23
albumentations==1.4.24
alive-progress==3.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
archspec @ file:///home/task_175812491784513/conda-bld/archspec_1758124989039/work
args==0.1.0
async-timeout==5.0.1
attrs==25.3.0
azure-core==1.38.0
azure-identity==1.25.1
azure-storage-blob==12.28.0
azure-storage-file-datalake==12.23.0
bcrypt==5.0.0
beautifulsoup4==4.13.5
boltons @ file:///croot/boltons_1751383775419/work
boto3==1.42.41
botocore==1.42.41
braceexpand==0.1.7
brotli==1.2.0
brotlicffi @ file:///croot/brotlicffi_1736182461069/work
bs4==0.0.2
cachetools==6.2.1
catalogue==2.0.10
certifi @ file:///croot/certifi_1754570635119/work/certifi
cffi==2.0.0
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
circuitbreaker==2.1.3
clean-fid==0.1.35
click==8.2.1
clint==0.5.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
conda @ file:///croot/conda_1722004606466/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1742218810516/work/src
conda-package-handling @ file:///croot/conda-package-handling_1731369017509/work
conda_package_streaming @ file:///croot/conda-package-streaming_1751307506088/work
contourpy==1.3.2
cramjam==2.11.0
cryptography==45.0.7
cycler==0.12.1
dataclasses==0.6
datasets==4.2.0
deepspeed==0.17.6
diffusers==0.32.2
dill==0.4.0
distro @ file:///croot/distro_1714488253808/work
einops==0.8.1
exceptiongroup==1.3.0
ExifRead-nocycle==3.0.1
filelock==3.19.1
fire==0.7.1
fonttools==4.59.2
frozendict @ file:///croot/frozendict_1713194832637/work
frozenlist==1.8.0
fsspec==2025.7.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.26.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.5.0
google-cloud-storage==2.10.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.70.0
graphemeu==0.7.2
grpcio==1.74.0
h11==0.16.0
h5py==3.12.1
hf-xet==1.1.9
hjson==3.1.0
hpsv2==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface_hub==0.36.1
idna==3.10
imageio==2.37.0
img2dataset==1.47.0
importlib_metadata==8.7.1
iniconfig==2.1.0
isodate==0.7.2
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.2
jsonlines==4.0.0
jsonpatch @ file:///croot/jsonpatch_1714483231291/work
jsonpointer @ file:///croot/jsonpointer_1753788433671/work
kiwisolver==1.4.9
lazy_loader==0.4
libmambapy @ file:///croot/mamba-split_1734469461757/work/libmambapy
lightning-utilities==0.15.2
littleutils==0.2.4
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.10.5
menuinst @ file:///croot/menuinst_1753464582829/work
mosaicml-streaming==0.11.0
mpmath==1.3.0
msal==1.34.0
msal-extensions==1.3.1
msgpack==1.1.1
multidict==6.7.0
multiprocess==0.70.16
networkx==3.4.2
ninja==1.13.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
oauth2client==4.1.3
oauthlib==3.3.1
oci==2.167.0
ogb==1.3.6
omegaconf==2.3.0
open_clip_torch==3.1.0
opencv-python-headless==4.11.0.86
outdated==0.2.2
packaging @ file:///croot/packaging_1753775359799/work
pandas==2.2.3
paramiko==3.5.1
pathlib==1.0.1
peft==0.18.1
pigz==2.9rc3
pillow==10.4.0
platformdirs @ file:///croot/platformdirs_1744273042065/work
pluggy @ file:///croot/pluggy_1733169602837/work
propcache==0.4.1
proto-plus==1.26.1
protobuf==3.20.3
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.11
pycosat @ file:///croot/pycosat_1736868416091/work
pycparser @ file:///home/task_175749556993857/conda-bld/pycparser_1757496041120/work
pydantic==2.11.7
pydantic_core==2.33.2
PyDrive==1.3.1
PyJWT==2.11.0
PyNaCl==1.6.2
pyOpenSSL==25.1.0
pyparsing==3.2.3
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
pytest==7.2.0
pytest-split==0.8.0
python-dateutil==2.9.0.post0
python-snappy==0.7.3
pytorch-triton-rocm==3.4.0+gitf7888497
pytz==2025.2
PyYAML==6.0.2
regex==2024.9.11
requests @ file:///croot/requests_1756709366904/work
requests-oauthlib==2.0.0
rsa==4.9.1
ruamel.yaml @ file:///croot/ruamel.yaml_1745960305322/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1745937152469/work
s3transfer==0.16.0
safetensors==0.4.5
SAM-2==1.0
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.15.3
seaborn==0.13.2
sentencepiece==0.2.0
sentry-sdk==2.35.1
simsimd==6.5.1
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
stringzilla==3.12.6
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2023.9.26
tiktoken==0.7.0
timm==0.9.1
tokenizers==0.20.3
tomli==2.2.1
torch==2.9.0.dev20250831+rocm6.4
torch-fidelity==0.3.0
torchaudio==2.8.0.dev20250901+rocm6.4
torchmetrics==1.8.2
torchvision==0.24.0.dev20250901+rocm6.4
tqdm @ file:///croot/tqdm_1738943501192/work
transformers==4.45.2
triton==3.0.0
truststore @ file:///croot/truststore_1753907171582/work
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3 @ file:///croot/urllib3_1750775463400/work
wandb==0.21.2
wcwidth==0.2.13
webdataset==0.2.111
Werkzeug==3.1.3
xformers==0.0.27.post2
-e git+https://github.com/selinyldrm/LANTERN.git@82517e03ea1e40aa12410f042dda23c330754c4e#egg=xllmx&subdirectory=models/base_models/lumina_mgpt
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard @ file:///croot/zstandard_1731356346222/work
zstd==1.5.7.3
